% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Grad}
\alias{Grad}
\title{Gradient for OT (Internal C function)}
\usage{
Grad(Mat, psi)
}
\arguments{
\item{Mat}{the cost matrix}

\item{psi}{the vector of potentials}
}
\description{
Gradient for OT (Internal C function)
}
