% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Objective}
\alias{Objective}
\title{Objective for OT (Internal C function)}
\usage{
Objective(Mat, psi)
}
\arguments{
\item{Mat}{the cost matrix}

\item{psi}{the vector of potentials}
}
\description{
Objective for OT (Internal C function)
}
