% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Plot}
\alias{Plot}
\title{Algorithm for plotting results of OT (Internal C function)}
\usage{
Plot(Mat, psi)
}
\arguments{
\item{Mat}{the cost matrix}

\item{psi}{the vector of potentials}
}
\description{
Algorithm for plotting results of OT (Internal C function)
}
