% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SGD_OT}
\alias{SGD_OT}
\title{SGD Algorithm for OT (Internal C function)}
\usage{
SGD_OT(C, Mat)
}
\arguments{
\item{C}{the descent constant}

\item{Mat}{the cost matrix}
}
\description{
SGD Algorithm for OT (Internal C function)
}
