% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{WassersteinDist}
\alias{WassersteinDist}
\title{Computes the Wassertein distance between a target and a reference}
\usage{
WassersteinDist(points, Ref, p = 2)
}
\arguments{
\item{points}{input points}

\item{Ref}{points from the reference,continuous measure}

\item{p}{1= W1, 2=W2}
}
\value{
Wasserstein distance
}
\description{
Computes the Wassertein distance between a target and a reference
}
