% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{WassersteinDistGaus}
\alias{WassersteinDistGaus}
\title{Computes the Wassertein distance to an isotropic Gaussian}
\usage{
WassersteinDistGaus(points, B, p = 2)
}
\arguments{
\item{points}{input points}

\item{B}{number of Monte-Carlo samples}

\item{p}{1= W1, 2=W2}
}
\value{
Wasserstein distance
}
\description{
Computes the Wassertein distance to an isotropic Gaussian
}
